Update the py-compile script (works with Python 3.12)
The version of automake's py-compile that lvm2 carries around in
autoconf/ is ancient. Particularly, it assumes the 'imp' module
exists, which as of Python 3.12 is no longer a safe assumption.
This just updates lvm2's copy of the file to the latest version
in upstream automake -
https://raw.githubusercontent.com/autotools-mirror/automake/4bd4640e03670e267d9763d377f107c09b4f2202/lib/py-compile
. From a quick look at the changes, the interface shouldn't have
changed, this should just work better with modern Python.

Signed-off-by: Adam Williamson <awilliam@redhat.com>